@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -15,36 +15,33 @@ package org.workhorse.activity;
 
 import org.workhorse.Recordable;
 import org.workhorse.actor.User;
-import org.workhorse.exec.ctx.LocalContext;
 
-import java.util.Date;
+import java.time.LocalDateTime;
 
 /**
  *  Tasks are activities in a process that require user interaction.
  *
  * @author Brennan Spies
  */
-public interface Task extends Activity, Recordable
+public interface Task extends AtomicActivity, Recordable
 {	
 	/**
 	 * Returns the user responsible for performing the task, or null
 	 * if not assigned.
 	 * @return The responsible user
 	 */
-	public User getPerformer();
-	
-	/**
-	 * Returns the date that work on the task was started (i.e., acquired
-	 * by the responsible <code>User</code>).
-	 * @return The task start date
-	 */
-	public Date getStartDate();
+	User getPerformer();
 	
 	/**
 	 * Returns the status of this task.
 	 * @return The status
 	 */
-	public Status getStatus();
+	Status getStatus();
 
-    public LocalContext getContext();
+	/**
+	 * Returns the date that this task is actually started by
+	 * the assigned user.
+	 * @return The date of task start
+     */
+	LocalDateTime getStartDate();
 }
