@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -20,8 +20,8 @@ package org.workhorse;
  * @author Brennan Spies
  */
 public interface Cancellable {
-	/**
-	 * Cancels the progress.
-	 */
-	public void cancel();
+    /**
+     * Cancels the progress.
+     */
+    void cancel();
 }
