@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -19,8 +19,8 @@ package org.workhorse;
  * @author Brennan Spies
  */
 public interface Startable {
-	/**
-	 * Starts the {@code Startable} instance.
-	 */
-	public void start();
+    /**
+     * Starts the {@code Startable} instance.
+     */
+    void start();
 }
