@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -13,15 +13,33 @@
  */
 package org.workhorse.activity;
 
-import org.workhorse.Identifiable;
-import org.workhorse.exec.Executable;
+import org.workhorse.Describable;
+import org.workhorse.exec.Contextual;
+import org.workhorse.exec.Execution;
+import org.workhorse.exec.RunState;
+import org.workhorse.exec.ctx.ReadValues;
+import org.workhorse.id.Identifiable;
+
+import java.util.UUID;
 
 /**
  * Base interface for all process activities.
  *
  * @author Brennan Spies
  */
-public interface Activity extends Identifiable<Long>, Executable
-{
+public interface Activity extends Identifiable<UUID>, Contextual, Describable {
 
+    /**
+     * Performs the activity.
+     * @param execution The current execution
+     */
+    RunState perform(Execution execution);
+
+    /**
+     * Gets the snapshot values as they were read from
+     * the parent context (at the time of input set
+     * evaluation).
+     * @return The read values
+     */
+    ReadValues getReadParentValues();
 }
