@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -13,7 +13,7 @@
  */
 package org.workhorse.actor;
 
-import org.workhorse.Identifiable;
+import org.workhorse.id.Identifiable;
 
 /**
  * @author Brennan Spies
@@ -25,5 +25,5 @@ public interface Actor extends Identifiable<Long>
 	 * Returns the name of the actor.
 	 * @return The name
 	 */
-	public String getName();
+	String getName();
 }
