@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -13,13 +13,14 @@
  */
 package org.workhorse.event;
 
-import org.workhorse.graph.node.Node;
+import org.workhorse.graph.Node;
 
 /**
+ * Abstract base class for {@link Event} implementations.
+ *
  * @author Brennan Spies
- * <p>Abstract base class for {@link Event} implementations.
  */
-public abstract class BaseEvent implements Event 
+public abstract class BaseEvent implements Event
 {
 	private Node source;
 	
@@ -31,10 +32,8 @@ public abstract class BaseEvent implements Event
 		this.source = source;
 	}
 
-	/**
-	 * @see org.workhorse.event.ThrownEvent#getSource()
-	 */
-	public Node getSource() {
+	/** {@inheritDoc} */
+	@Override public Node getSource() {
 		return source;
 	}
 }
