@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -16,6 +16,8 @@ package org.workhorse.event;
 import org.workhorse.expr.Condition;
 
 /**
+ * Represents a conditional event.
+ *
  * @author Brennan Spies
  */
 public interface Conditional extends Event, Condition {}
