@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -13,9 +13,8 @@
  */
 package org.workhorse.event;
 
-import org.workhorse.graph.node.Node;
-
 import com.google.inject.TypeLiteral;
+import org.workhorse.graph.Node;
 
 /**
  * An event indicating that an error has occurred.
@@ -24,7 +23,7 @@ import com.google.inject.TypeLiteral;
  */
 public class Error extends BaseEvent implements ThrownEvent<Error>
 {
-	public static final EventType<Error> type = 
+	public static final EventType<Error> type =
 		new EventType<Error>(new TypeLiteral<Error>(){});
 
 	private final String name;
