@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -13,7 +13,7 @@
  */
 package org.workhorse.event;
 
-import org.workhorse.graph.node.Node;
+import org.workhorse.graph.Node;
 
 /**
  * A termination event.
@@ -24,7 +24,6 @@ public class Termination extends BaseEvent
 {
 	/**
 	 * Default constructor.
-     *
 	 * @param source The source node
 	 */
 	public Termination(Node source) {
