@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -13,21 +13,29 @@
  */
 package org.workhorse.event;
 
-import org.workhorse.graph.node.Node;
+import org.workhorse.graph.Node;
 
 /**
+ * Represents a timed wait event.
+ *
  * @author Brennan Spies
  */
 public class TimedWait extends BaseEvent
 {
 	private long interval;
-	
+
+	/**
+	 * Creates the timed wait event.
+	 * @param source The source node
+	 * @param interval The interval of the wait
+     */
 	public TimedWait(Node source, long interval) {
 		super(source);
 		this.interval = interval;
 	}
 	
 	/**
+	 * Returns the wait interval.
 	 * @return The wait interval
 	 */
 	public long getInterval() { return interval; }
