@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -27,5 +27,5 @@ public interface Trigger<E extends ThrownEvent<E>> {
 	 * @param context The execution context
 	 * @return The event
 	 */
-	public E generate(Context context);
+	E generate(Context context);
 }
