@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -48,7 +48,7 @@ public class ErrorCatcher implements Catcher<Error>
 	 * {@inheritDoc}
 	 */
 	public void handle(Error error, Execution execution) {
-		//TODO
+		//TODO implement
 	}
 
 	/**
