@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -21,12 +21,12 @@ import org.workhorse.exec.Execution;
  *
  * @author Brennan Spies
  */
-public interface Handler<E extends Event> 
+public interface Handler<E extends Event>
 {
 	/**
 	 * Method called when an event is received by an {@link Execution}.
 	 * @param event The event
 	 * @param execution The execution processing the event
 	 */
-	public void handle(E event, Execution execution);
+	void handle(E event, Execution execution);
 }
