@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -14,7 +14,7 @@
 package org.workhorse.graph;
 
 /**
- * Represents a swimlane in a BPMN diagram.
+ * Represents a swim lane in a BPMN diagram.
  *
  * @author Brennan Spies
  */
