@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -14,9 +14,9 @@
 package org.workhorse.process;
 
 /**
+ * Enumeration for process states.
+ *
  * @author Brennan Spies
- * 
- * <p>Enumeration for process states.</p>
  */
 public enum State {
 	READY,
